pyproject.toml
[project]
name = "aiod_enhanced_interaction"
description = "An Enhanced Asset Search for AI on Demand"
version = "1.0.20250228"
requires-python = ">=3.11"
authors = [
    { name = "Marcel Veselý", email = "marcel.vesely@kinit.sk" },
    { name = "Alexander Brecko", email = "alexander.brecko@kinit.sk" },
    { name = "Andrej Ridzik", email = "andrej.ridzik@kinit.sk" }
]
dependencies = [
    "fastapi==0.115.12",
    "uvicorn==0.34.0",
    "numpy==2.2.4",
    "pandas==2.2.3",
    "pydantic==2.11.9",
    "pydantic-settings==2.8.1",
    "pydantic-ai-slim[openai,logfire]==1.11.0",
    "tqdm==4.67.1",
    "python-dotenv==1.1.0",
    "pymilvus==2.5.5",
    "transformers==4.44.2",
    "sentence-transformers==3.0.1",
    "apscheduler==3.10.4",
    "beanie==1.29.0",
    "motor==3.7.1",
    "pymongo==4.13.0",
    "tenacity==9.1.2",
    "Crawl4AI==0.7.2",
    "beautifulsoup4==4.13.4",
    "mistralai==1.9.2",
    "nltk==3.9.1",
    "ollama==0.4.7"
]
readme = "README.md"

[project.optional-dependencies]
dev = [
    "black==24.8.0",
    "flake8==7.2.0",
    "isort==6.0.1",
    "pre-commit==3.3.3",
    "torch==2.6.0",
    "jupyter_core==5.8.1",
    "tornado>=6.2",
    "pyzmq>=23.0",
    "ipykernel==7.1.0"

]

[tool.setuptools]
py-modules = []

[tool.ruff]
line-length = 100
exclude = [
    "experiments",
]

[tool.codespell]
ignore-words-list = "checkin"

[tool.ruff.lint]
select = [
    "S",
    "I",
]

[tool.ruff.lint.isort]
known-local-folder = ["app"]
